Allow selectively compiling of micropython native compressor/decompressor
